+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+ * Add slug generator interface. [#11]
 
 ## [1.0.3] (2018-11-23)
 
@@ -40,4 +41,5 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 [1.0.0-alpha1]: https://github.com/ausi/slug-generator/commits/v1.0.0-alpha1
 
 [#13]: https://github.com/ausi/slug-generator/issues/13
+[#11]: https://github.com/ausi/slug-generator/issues/11
 [#9]: https://github.com/ausi/slug-generator/issues/9